refactor: util함수 정리, 토큰 디코딩 함수 추가 - jwtUtils.ts 파일 생성   - 기존 관련 함수를 해당 파일로 이동   - decodeJWT함수 추가: 토큰의 role과 hasProfile정보를 디코딩하여 반환 - getUserFromRequest함수에 token으로부터 user정보( role과 hasProfile)를 얻는 코드 추가(현재는 주석 처리, 백엔드 수정이 되면 적용해볼 것)
